Look for repository when open file
