+++ feature-git/api/src/main/kotlin/com/blacksquircle/ui/feature/git/api/interactor/GitInteractor.kt
@@ -21,8 +21,6 @@ import kotlinx.coroutines.flow.Flow
 
 interface GitInteractor {
 
-    suspend fun getRepoPath(fileModel: FileModel): String
-
     suspend fun cloneRepository(
         fileModel: FileModel,
         url: String,
